+++ StratifiedEventQueue/States/Gates/Or.cs
@@ -26,7 +26,8 @@ namespace StratifiedEventQueue.States.Gates
         /// <param name="riseDelay">The rise delay.</param>
         /// <param name="fallDelay">The fall delay.</param>
         /// <exception cref="ArgumentNullException">Thrown if <paramref name="a"/>, <paramref name="b"/>, <paramref name="name"/> or <paramref name="outputName"/> is <c>null</c>.</exception>
-        public Or(string name, string outputName, IState<Signal> a, IState<Signal> b, ulong riseDelay = 0, ulong fallDelay = 0)
+        public Or(string name, string outputName, IState<Signal> a, IState<Signal> b,
+            uint riseDelay = 0, uint fallDelay = 0)
             : base(name, outputName, riseDelay, fallDelay)
         {
             A = a ?? throw new ArgumentNullException(nameof(a));
